move delta `Node` into `resolve.rs`
The `Node` type is only used in `resolve.rs`, and using it safely
depends on using the child items correctly, so let's define it there.
